Move notification source.
