+++ pkg/plugins/datalayer/notificationsource/plugin.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package datalayer
+package notificationsource
 
 import (
 	"context"
@@ -26,45 +26,45 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/llm-d/llm-d-inference-payload-processor/pkg/framework"
+	"github.com/llm-d/llm-d-inference-payload-processor/pkg/framework/datalayer"
 )
 
 const (
-	NotificationSourcePluginType = "notification-source"
+	PluginType = "notification-source"
 
 	defaultBufferSize = 10000
 )
 
 // compile-time interface assertion
-var _ framework.NotificationSource = &notificationSource{}
+var _ datalayer.NotificationSource = &notificationSource{}
 
 type notificationSource struct {
 	name       framework.TypedName
-	ch         chan framework.Event
-	extractors []framework.Extractor
+	ch         chan datalayer.Event
+	extractors []datalayer.Extractor
 
 	started atomic.Bool
 	cancel  context.CancelFunc
 	done    chan struct{}
 }
 
-// NotificationSourceFactory is the factory function for NotificationSource.
-func NotificationSourceFactory(name string, _ json.RawMessage, _ framework.Handle) (framework.Plugin, error) {
-	src, err := NewNotificationSource(name)
+// Factory is the factory function for NotificationSource.
+func Factory(name string, _ json.RawMessage, _ framework.Handle) (framework.Plugin, error) {
+	src, err := New(name)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create '%s' plugin - %w", NotificationSourcePluginType, err)
+		return nil, fmt.Errorf("failed to create '%s' plugin - %w", PluginType, err)
 	}
 	return src, nil
 }
 
-// NewNotificationSource creates a NotificationSource that delivers each event
-// to the given extractors as it arrives.
-func NewNotificationSource(name string, extractors ...framework.Extractor) (framework.NotificationSource, error) {
+// New creates a NotificationSource that delivers each event to the given extractors as it arrives.
+func New(name string, extractors ...datalayer.Extractor) (datalayer.NotificationSource, error) {
 	if name == "" {
-		return nil, fmt.Errorf("name is required for plugin '%s'", NotificationSourcePluginType)
+		return nil, fmt.Errorf("name is required for plugin '%s'", PluginType)
 	}
 	return &notificationSource{
-		name:       framework.TypedName{Type: NotificationSourcePluginType, Name: name},
-		ch:         make(chan framework.Event, defaultBufferSize),
+		name:       framework.TypedName{Type: PluginType, Name: name},
+		ch:         make(chan datalayer.Event, defaultBufferSize),
 		extractors: extractors,
 		done:       make(chan struct{}),
 	}, nil
@@ -72,12 +72,12 @@ func NewNotificationSource(name string, extractors ...framework.Extractor) (fram
 
 func (n *notificationSource) TypedName() framework.TypedName { return n.name }
 
-func (n *notificationSource) RegisterExtractor(e framework.Extractor) {
+func (n *notificationSource) RegisterExtractor(e datalayer.Extractor) {
 	n.extractors = append(n.extractors, e)
 }
 
 // Notify fires an event non-blocking; drops silently if the buffer is full.
-func (n *notificationSource) Notify(e framework.Event) {
+func (n *notificationSource) Notify(e datalayer.Event) {
 	select {
 	case n.ch <- e:
 	default:
@@ -118,7 +118,7 @@ func (n *notificationSource) eventLoop(ctx context.Context, ready chan struct{})
 			// Extractors are called sequentially; current extractors are in-memory only.
 			// Switch to a WaitGroup if any extractor performs I/O.
 			for _, ext := range n.extractors {
-				if err := ext.Extract(ctx, []framework.Event{e}); err != nil {
+				if err := ext.Extract(ctx, []datalayer.Event{e}); err != nil {
 					logger.Error(err, "extractor error", "extractor", ext.TypedName())
 				}
 			}
